pyproject.toml
[tool.poetry]
name = "vladimir-putin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "vladimir_putin"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
disnake = "^2.9.0"
schedule = "^1.2.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
